requirements.txt
notebook==7.4.1
numpy==2.2.5
pandas==2.2.3
pydantic==2.11.3
genson==1.3.0
fastapi==0.115.12
uvicorn==0.34.2
PyMuPDF==1.25.5
pillow==11.2.1
torch==2.7.0
torchvision==0.22.0
transformers==4.51.3
accelerate==1.6.0
bitsandbytes==0.45.1
huggingface_hub==0.30.2